+++ packages/suite/src/actions/firmware/__tests__/firmwareActions.test.ts
