+++ packages/suite/src/actions/firmware/constants/firmware.ts
