+++ packages/suite/src/actions/firmware/constants/index.ts
@@ -1,0 +1,3 @@
+import * as FIRMWARE from './firmware';
+
+export { FIRMWARE };
