+++ packages/suite/src/actions/settings/constants/index.ts
@@ -1,6 +1,5 @@
 import * as WALLET_SETTINGS from './walletSettings';
 import * as DEVICE_SETTINGS from './deviceSettings';
-import * as FIRMWARE from './firmware';
 import * as RECOVERY from './recovery';
 
-export { WALLET_SETTINGS, DEVICE_SETTINGS, FIRMWARE, RECOVERY };
+export { WALLET_SETTINGS, DEVICE_SETTINGS, RECOVERY };
