+++ packages/suite/src/middlewares/suite/index.ts
@@ -1,8 +1,7 @@
 import log from './logMiddleware';
 import suite from './suiteMiddleware';
 import redirect from './redirectMiddleware';
-import firmware from './firmwareMiddleware';
 import analytics from './analyticsMiddleware';
 import buttonRequest from './buttonRequestMiddleware';
 
-export default [log, redirect, suite, firmware, analytics, buttonRequest];
+export default [log, redirect, suite, analytics, buttonRequest];
