+++ packages/suite/src/types/firmware/index.ts
@@ -1,0 +1,4 @@
+import { FirmwareActions as FirmwareActions$ } from '@firmware-actions/firmwareActions';
+
+export type FirmwareActions = FirmwareActions$;
+
