+++ packages/suite/src/types/settings/index.ts
@@ -1,7 +1,6 @@
-import { FirmwareActions } from '@settings-actions/firmwareActions';
 import { WalletSettingsActions } from '@settings-actions/walletSettingsActions';
 import { RecoveryActions } from '@settings-actions/recoveryActions';
 
-export type SettingsActions = FirmwareActions | WalletSettingsActions | RecoveryActions;
+export type SettingsActions = WalletSettingsActions | RecoveryActions;
 
 export type WordCount = 12 | 18 | 24;
